Fix another issue with new instances

+++ scripts/stop_container.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-cd /app
-/usr/local/bin/docker-compose --file docker-compose-production.yml down -d > /var/log/docker-compose.log 2> /var/log/docker-compose.log
-
-# Always succeed; we may be running on a new instance with no containers
-exit 0
+if [ -d /app ]
+then
+  cd /app
+  /usr/local/bin/docker-compose --file docker-compose-production.yml down -d > /var/log/docker-compose.log 2> /var/log/docker-compose.log
+fi
